requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pydantic-settings
python-dotenv
convex
